Fix segfault on schema change after an error in change
In case of error in change the pointer to the hash table to invalidate
is not reset as the commit handler is not called, bug the memory context
is deleted. At next schema change the handler will try to access invalid
memory and boom.

The problem is fixed resetting and restoring the schema callback during
the change handler. The extra care taken (allocating the hash table in
the context of the plugin, deleting it and clearing the invalidation
pointer on cleanup) are not useful to fix the bug (the cleanup is
skipped too on error) but I think they don't hurt.