Python source and tests split
The python package is no more in the root directory: use:

    python setup.py develop

to make it importable. Tests can now run using

    python setup.py test

as well as running 'pytest'. Note though that the first form will always
test the content of the project directory, whereas the second can be
used to test the installed package (if you are not in develop mode).